@@ -1,12 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <RootNamespace>Moq.AutoMock.Tests</RootNamespace>
     <AssemblyName>Moq.AutoMock.Tests</AssemblyName>
-    <Copyright>Copyright © 2016</Copyright>
+    <Copyright>Copyright © 2020</Copyright>
 
     <IsPackable>false</IsPackable>
+    <NoWarn>$(NoWarn),CA1707</NoWarn>
   </PropertyGroup>
   
   <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
@@ -23,11 +24,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Moq" version="4.8.2" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.4.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
-    <ProjectReference Include="..\Moq.AutoMock\Moq.AutoMock.csproj" />
+    <Compile Include="..\System.Diagnostics.CodeAnalysis.cs" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Moq" version="4.13.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.1.0" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.1.0" />
+    <ProjectReference Include="..\Moq.AutoMock\Moq.AutoMock.csproj" />
+  </ItemGroup>
 </Project>
