@@ -1,4 +1,5 @@
 ﻿using Moq.AutoMock.Resolvers;
+using Moq.Language;
 using Moq.Language.Flow;
 using System;
 using System.Collections.Generic;
@@ -6,7 +7,6 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Runtime.ExceptionServices;
-using Moq.Language;
 
 namespace Moq.AutoMock
 {
@@ -15,23 +15,35 @@ namespace Moq.AutoMock
     /// </summary>
     public partial class AutoMocker
     {
-        private readonly Dictionary<Type, IInstance> typeMap = new Dictionary<Type, IInstance>();
+        private readonly Dictionary<Type, IInstance> _typeMap = new Dictionary<Type, IInstance>();
 
+        /// <summary>
+        /// Creates a new instance of AutoMocker.
+        /// </summary>
         public AutoMocker()
             : this(MockBehavior.Default)
         {
         }
 
+        /// <summary>
+        /// Creates a new instance of AutoMocker.
+        /// </summary>
         public AutoMocker(MockBehavior mockBehavior)
             : this(mockBehavior, DefaultValue.Empty)
         {
         }
 
+        /// <summary>
+        /// Creates a new instance of AutoMocker.
+        /// </summary>
         public AutoMocker(MockBehavior mockBehavior, DefaultValue defaultValue)
             : this(mockBehavior, defaultValue, callBase: false)
         {
         }
 
+        /// <summary>
+        /// Creates a new instance of AutoMocker.
+        /// </summary>
         public AutoMocker(MockBehavior mockBehavior, DefaultValue defaultValue, bool callBase)
         {
             MockBehavior = mockBehavior;
@@ -46,9 +58,22 @@ namespace Moq.AutoMock
             };
         }
 
+        /// <summary>
+        /// Gets the behavior of created Mocks.
+        /// </summary>
         public MockBehavior MockBehavior { get; }
+
+        /// <summary>
+        /// Specifies the behavior to use when returning default values for unexpected invocations on loose mocks.
+        /// </summary>
         public DefaultValue DefaultValue { get; }
+        /// <summary>
+        /// Whether the base member virtual implementation will be called for mocked classes if no setup is matched.
+        /// </summary>
         public bool CallBase { get; }
+        /// <summary>
+        /// A collection of resolvers that are used to resolve services.
+        /// </summary>
         public ICollection<IMockResolver> Resolvers { get; }
 
         private IInstance Resolve(Type serviceType)
@@ -57,21 +82,21 @@ namespace Moq.AutoMock
             {
                 Type elmType = serviceType.GetElementType();
                 MockArrayInstance instance = new MockArrayInstance(elmType);
-                if (typeMap.TryGetValue(elmType, out var element))
+                if (_typeMap.TryGetValue(elmType, out var element))
                     instance.Add(element);
                 return instance;
             }
 
-            var resolved = Resolve(serviceType, null);
-            switch (resolved)
+            object? resolved = Resolve(serviceType, null);
+            return resolved switch
             {
-                case Mock mock: return new MockInstance(mock);
-                case IInstance instance: return instance;
-                case object _: return new RealInstance(resolved);
-                default: return null;
-            }
+                Mock mock => new MockInstance(mock),
+                IInstance instance => instance,
+                _ => new RealInstance(resolved),
+            };
         }
-        private object Resolve(Type serviceType, object initialValue)
+
+        private object? Resolve(Type serviceType, object? initialValue)
         {
             var context = new MockResolutionContext(this, serviceType, initialValue);
 
@@ -129,11 +154,16 @@ namespace Moq.AutoMock
         /// setup in the container.</returns>
         public object CreateInstance(Type type, bool enablePrivate)
         {
-            var bindingFlags = GetBindingFlags(enablePrivate);
-            var arguments = CreateArguments(type, bindingFlags);
+            if (type is null) throw new ArgumentNullException(nameof(type));
+
+            BindingFlags bindingFlags = GetBindingFlags(enablePrivate);
+            object?[] arguments = CreateArguments(type, bindingFlags);
             try
             {
-                var ctor = type.SelectCtor(typeMap.Keys.ToArray(), bindingFlags);
+                var ctor = type.SelectCtor(_typeMap.Keys.ToArray(), bindingFlags);
+                if (ctor is null)
+                    throw new ArgumentException($"`{type}` does not have an acceptable constructor.", nameof(type));
+
                 return ctor.Invoke(arguments);
             }
             catch (TargetInvocationException e)
@@ -151,7 +181,7 @@ namespace Moq.AutoMock
         /// </summary>
         /// <typeparam name="T">The instance that you want to build</typeparam>
         /// <returns>An instance with virtual and abstract members mocked</returns>
-        public T CreateSelfMock<T>() where T : class
+        public T CreateSelfMock<T>() where T : class?
             => CreateSelfMock<T>(false);
 
         /// <summary>
@@ -164,7 +194,7 @@ namespace Moq.AutoMock
         /// <param name="enablePrivate">When true, private constructors will also be used to
         /// create mocks.</param>
         /// <returns>An instance with virtual and abstract members mocked</returns>
-        public T CreateSelfMock<T>(bool enablePrivate) where T : class
+        public T CreateSelfMock<T>(bool enablePrivate) where T : class?
         {
             var arguments = CreateArguments(typeof(T), GetBindingFlags(enablePrivate));
 
@@ -173,8 +203,12 @@ namespace Moq.AutoMock
                 DefaultValue = DefaultValue,
                 CallBase = CallBase
             };
+
             var resolved = Resolve(typeof(T), mock);
-            return (resolved as Mock<T>)?.Object;
+            if (resolved is Mock<T> m)
+                return m.Object;
+            
+            return default(T)!;
         }
 
         #endregion Create Instance/SelfMock
@@ -193,12 +227,12 @@ namespace Moq.AutoMock
         /// </summary>
         /// <param name="type">The type of service to use</param>
         /// <param name="service">The service to use</param>
-        public void Use(Type type, object service)
+        public void Use(Type type, object? service)
         {
-            if (type == null) throw new ArgumentNullException(nameof(type));
+            if (type is null) throw new ArgumentNullException(nameof(type));
             if (service != null && !type.IsInstanceOfType(service))
                 throw new ArgumentException($"{nameof(service)} is not of type {type}");
-            typeMap[type] = new RealInstance(service);
+            _typeMap[type] = new RealInstance(service);
         }
 
         /// <summary>
@@ -209,7 +243,7 @@ namespace Moq.AutoMock
         public void Use<TService>(Mock<TService> mockedService)
             where TService : class
         {
-            typeMap[typeof(TService)] = new MockInstance(mockedService);
+            _typeMap[typeof(TService)] = new MockInstance(mockedService ?? throw new ArgumentNullException(nameof(mockedService)));
         }
 
         /// <summary>
@@ -220,6 +254,8 @@ namespace Moq.AutoMock
         public void Use<TService>(Expression<Func<TService, bool>> setup)
             where TService : class
         {
+            if (setup is null) throw new ArgumentNullException(nameof(setup));
+
             Use(Mock.Get(Mock.Of(setup)));
         }
 
@@ -233,7 +269,8 @@ namespace Moq.AutoMock
         /// </summary>
         /// <typeparam name="TService">The class or interface to search on</typeparam>
         /// <returns>The object that implements TService</returns>
-        public TService Get<TService>() => (TService)Get(typeof(TService));
+        public TService? Get<TService>() where TService : class
+            => Get(typeof(TService)) is TService service ? service : null;
 
         /// <summary>
         /// Searches and retrieves an object from the container that matches the serviceType. This can be
@@ -241,12 +278,14 @@ namespace Moq.AutoMock
         /// </summary>
         /// <param name="serviceType">The type of service to retrieve</param>
         /// <returns></returns>
-        public object Get(Type serviceType)
+        public object? Get(Type serviceType)
         {
-            if (!typeMap.TryGetValue(serviceType, out var instance) || instance is null)
-                instance = typeMap[serviceType] = Resolve(serviceType);
+            if (serviceType is null) throw new ArgumentNullException(nameof(serviceType));
 
-            if (instance == null)
+            if (!_typeMap.TryGetValue(serviceType, out var instance) || instance is null)
+                instance = _typeMap[serviceType] = Resolve(serviceType);
+
+            if (instance is null)
                 throw new ArgumentException($"{serviceType} could not resolve to an object.", nameof(serviceType));
             return instance.Value;
         }
@@ -271,18 +310,18 @@ namespace Moq.AutoMock
         /// <returns>A mock of serviceType</returns>
         public Mock GetMock(Type serviceType)
         {
-            if (serviceType == null) throw new ArgumentNullException(nameof(serviceType));
+            if (serviceType is null) throw new ArgumentNullException(nameof(serviceType));
 
             return GetMockImplementation(serviceType);
         }
 
         private Mock GetMockImplementation(Type serviceType)
         {
-            if (!typeMap.TryGetValue(serviceType, out var instance) || instance is null)
-                instance = typeMap[serviceType] = Resolve(serviceType);
+            if (!_typeMap.TryGetValue(serviceType, out var instance) || instance is null)
+                instance = _typeMap[serviceType] = Resolve(serviceType);
 
             if (instance == null || !instance.IsMock)
-                throw new ArgumentException($"Registered service `{Get(serviceType).GetType()}` was not a mock");
+                throw new ArgumentException($"Registered service `{Get(serviceType)?.GetType()}` was not a mock");
 
             var mockInstance = (MockInstance)instance;
             return mockInstance.Mock;
@@ -291,16 +330,6 @@ namespace Moq.AutoMock
         #endregion GetMock
 
         #region Setup
-
-        /// <summary>
-        /// Shortcut for mock.Setup(...), creating the mock when necessary.
-        /// </summary>
-        [Obsolete("No longer supported with Moq 4.8 and later. Use Setup<TService, TReturn> instead.", true)]
-        public ISetup<TService, object> Setup<TService>(Expression<Func<TService, object>> setup)
-            where TService : class
-        {
-            throw new NotSupportedException("No longer supported in Moq 4.8 and later. Use Setup<TService, TReturn> instead.");
-        }
 
         /// <summary>
         /// Shortcut for mock.Setup(...), creating the mock when necessary.
@@ -356,6 +385,8 @@ namespace Moq.AutoMock
         public ISetupSequentialResult<TReturn> SetupSequence<TService, TReturn>(Expression<Func<TService, TReturn>> setup)
             where TService : class
         {
+            if (setup is null) throw new ArgumentNullException(nameof(setup));
+
             return Setup<ISetupSequentialResult<TReturn>, TService>(m => m.SetupSequence(setup));
         }
 
@@ -371,14 +402,16 @@ namespace Moq.AutoMock
         /// </summary>
         public void Combine(Type type, params Type[] forwardTo)
         {
+            if (type is null) throw new ArgumentNullException(nameof(type));
+
             if (!(Resolve(type) is MockInstance mockObject))
                 throw new ArgumentException($"{type} did not resolve to a Mock", nameof(type));
 
             forwardTo.Aggregate(mockObject.Mock, As);
             foreach (var serviceType in forwardTo.Concat(new[] { type }))
-                typeMap[serviceType] = mockObject;
+                _typeMap[serviceType] = mockObject;
 
-            Mock As(Mock mock, Type forInterface)
+            static Mock As(Mock mock, Type forInterface)
             {
                 var method = mock.GetType().GetMethods().First(x => x.Name == nameof(Mock.As))
                     .MakeGenericMethod(forInterface);
@@ -395,7 +428,7 @@ namespace Moq.AutoMock
         /// </summary>
         public void VerifyAll()
         {
-            foreach (var pair in typeMap)
+            foreach (var pair in _typeMap)
             {
                 if (pair.Value is MockInstance instance)
                     instance.Mock.VerifyAll();
@@ -407,7 +440,7 @@ namespace Moq.AutoMock
         /// </summary>
         public void Verify()
         {
-            foreach (var pair in typeMap)
+            foreach (var pair in _typeMap)
             {
                 if (pair.Value is MockInstance instance)
                     instance.Mock.Verify();
@@ -423,6 +456,8 @@ namespace Moq.AutoMock
         public void Verify<T, TResult>(Expression<Func<T, TResult>> expression)
             where T : class
         {
+            if (expression is null) throw new ArgumentNullException(nameof(expression));
+
             var mock = GetMock<T>();
             mock.Verify(expression);
         }
@@ -437,6 +472,8 @@ namespace Moq.AutoMock
         public void Verify<T, TResult>(Expression<Func<T, TResult>> expression, Times times)
             where T : class
         {
+            if (expression is null) throw new ArgumentNullException(nameof(expression));
+
             var mock = GetMock<T>();
             mock.Verify(expression, times);
         }
@@ -451,6 +488,9 @@ namespace Moq.AutoMock
         public void Verify<T, TResult>(Expression<Func<T, TResult>> expression, Func<Times> times)
             where T : class
         {
+            if (expression is null) throw new ArgumentNullException(nameof(expression));
+            if (times is null) throw new ArgumentNullException(nameof(times));
+
             var mock = GetMock<T>();
 
             mock.Verify(expression, times);
@@ -463,9 +503,11 @@ namespace Moq.AutoMock
         /// <typeparam name="TResult">Return type of the full expression</typeparam>
         /// <param name="expression"></param>
         /// <param name="failMessage"></param>
-        public void Verify<T, TResult>(Expression<Func<T, TResult>> expression, String failMessage)
+        public void Verify<T, TResult>(Expression<Func<T, TResult>> expression, string failMessage)
             where T : class
         {
+            if (expression is null) throw new ArgumentNullException(nameof(expression));
+
             var mock = GetMock<T>();
             mock.Verify(expression, failMessage);
         }
@@ -478,9 +520,12 @@ namespace Moq.AutoMock
         /// <param name="expression"></param>
         /// <param name="times"></param>
         /// <param name="failMessage"></param>
-        public void Verify<T, TResult>(Expression<Func<T, TResult>> expression, Times times, String failMessage)
+        public void Verify<T, TResult>(Expression<Func<T, TResult>> expression, Times times, string failMessage)
             where T : class
         {
+            if (expression is null) throw new ArgumentNullException(nameof(expression));
+            if (failMessage is null) throw new ArgumentNullException(nameof(failMessage));
+
             var mock = GetMock<T>();
             mock.Verify(expression, times, failMessage);
         }
@@ -492,26 +537,28 @@ namespace Moq.AutoMock
         private static BindingFlags GetBindingFlags(bool enablePrivate)
         {
             var bindingFlags = BindingFlags.Instance | BindingFlags.Public;
-            if (enablePrivate) bindingFlags = bindingFlags | BindingFlags.NonPublic;
+            if (enablePrivate) bindingFlags |= BindingFlags.NonPublic;
             return bindingFlags;
         }
 
-        private object[] CreateArguments(Type type, BindingFlags bindingFlags)
+        private object?[] CreateArguments(Type type, BindingFlags bindingFlags)
         {
-            var ctor = type.SelectCtor(typeMap.Keys.ToArray(), bindingFlags);
-            var arguments = ctor.GetParameters().Select(x => Get(x.ParameterType)).ToArray();
-            return arguments;
+            ConstructorInfo ctor = type.SelectCtor(_typeMap.Keys.ToArray(), bindingFlags);
+            if (ctor is null)
+                throw new ArgumentException($"`{type}` does not have an acceptable constructor.", nameof(type));
+
+            return ctor.GetParameters().Select(x => Get(x.ParameterType)).ToArray();
         }
 
         private Mock GetOrMakeMockFor(Type type)
         {
-            if (!typeMap.TryGetValue(type, out var instance) || !instance.IsMock)
+            if (!_typeMap.TryGetValue(type, out var instance) || !instance.IsMock)
                 instance = Resolve(type);
 
             if (!(instance is MockInstance mockInstance))
                 throw new ArgumentException($"{type} does not resolve to a Mock");
 
-            typeMap[type] = mockInstance;
+            _typeMap[type] = mockInstance;
             return mockInstance.Mock;
         }
 
