@@ -17,7 +17,7 @@ namespace Moq.AutoMock.Tests
         public void It_creates_mock_from_interface_from_type_parameter()
         {
             var mocker = new AutoMocker();
-            Mock<IService1> mock = mocker.GetMock(typeof(IService1)) as Mock<IService1>;
+            Mock<IService1>? mock = mocker.GetMock(typeof(IService1)) as Mock<IService1>;
             Assert.IsNotNull(mock);
         }
 
@@ -25,7 +25,7 @@ namespace Moq.AutoMock.Tests
         public void It_creates_mock_from_class_from_generic_type_parameter()
         {
             var mocker = new AutoMocker();
-            Mock<Empty> mock = mocker.GetMock<Empty>();
+            Mock<Empty>? mock = mocker.GetMock<Empty>();
             Assert.IsNotNull(mock);
         }
 
@@ -33,7 +33,7 @@ namespace Moq.AutoMock.Tests
         public void It_creates_mock_from_class_from_type_parameter()
         {
             var mocker = new AutoMocker();
-            Mock<Empty> mock = mocker.GetMock(typeof(Empty)) as Mock<Empty>;
+            Mock<Empty>? mock = mocker.GetMock(typeof(Empty)) as Mock<Empty>;
             Assert.IsNotNull(mock);
         }
     }
