@@ -13,7 +13,7 @@ namespace Moq.AutoMock.Tests
             var mock = new Mock<IService2>();
             mocker.Use(mock);
             var name = mock.Object.Name;
-            mocker.Verify<IService2>(x => x.Name);
+            mocker.Verify<IService2>(x => x.Name!);
         }
 
         [TestMethod]
@@ -23,7 +23,7 @@ namespace Moq.AutoMock.Tests
             mocker.Setup<IServiceWithPrimitives, long>(s => s.ReturnsALong()).Returns(100L);
 
             var mock = mocker.Get<IServiceWithPrimitives>();
-            Assert.AreEqual(100L, mock.ReturnsALong());
+            Assert.AreEqual(100L, mock!.ReturnsALong());
 
             mocker.Verify<IServiceWithPrimitives, long>(s => s.ReturnsALong(), Times.Once());
         }
@@ -35,7 +35,7 @@ namespace Moq.AutoMock.Tests
             mocker.Setup<IService1>(x => x.Void()).Verifiable();
             mocker.Setup<IService5, string>(x => x.Name).Returns("Test");
 
-            mocker.Get<IService1>().Void();
+            mocker.Get<IService1>()!.Void();
 
             mocker.Verify();
         }
