@@ -8,9 +8,14 @@ namespace Moq.AutoMock.Tests
     [TestClass]
     public class ResolvesLazy
     {
-        [TestMethod] public void ResolvesLazyObjectFromContainer() => Resolves(new object());
-        [TestMethod] public void ResolvesLazyNumberFromContainer() => Resolves(42L);
-        [TestMethod] public void ResolvesLazyStringFromContainer() => Resolves(nameof(ResolvesLazy));
+        [TestMethod] 
+        public void ResolvesLazyObjectFromContainer() => Resolves(new object());
+        
+        [TestMethod] 
+        public void ResolvesLazyNumberFromContainer() => Resolves(42L);
+        
+        [TestMethod] 
+        public void ResolvesLazyStringFromContainer() => Resolves(nameof(ResolvesLazy));
 
         private static void Resolves<T>(T expected)
         {
@@ -19,7 +24,7 @@ namespace Moq.AutoMock.Tests
 
             var lazy = mocker.Get<Lazy<T>>();
             Assert.IsNotNull(lazy);
-            Assert.AreEqual(expected, lazy.Value);
+            Assert.AreEqual(expected, lazy!.Value);
         }
     }
 }
