@@ -1,6 +1,7 @@
 ﻿namespace Moq.AutoMock.Tests
 {
-    class WithArrayParameter
+#pragma warning disable CA1801, CA1812  //is an internal class that is apparently never instantiated
+    internal class WithArrayParameter
     {
         public WithArrayParameter() { }
         public WithArrayParameter(string[] array) { }
