@@ -1,6 +1,7 @@
 ﻿namespace Moq.AutoMock.Tests
 {
-    class WithSealedParameter
+#pragma warning disable CA1801, CA1812  //is an internal class that is apparently never instantiated
+    internal class WithSealedParameter
     {
         public WithSealedParameter() { }
         public WithSealedParameter(string @sealed) { }
