@@ -2,14 +2,15 @@
 {
     public class WithServiceInternal
     {
-        public IService1 Service { get; set; }
+        public IService1? Service { get; set; }
 
-        internal WithServiceInternal(IService1 service)
+        internal WithServiceInternal(IService1? service)
         {
             Service = service;
         }
 
-        public WithServiceInternal() : this(null)
+        public WithServiceInternal() 
+            : this(null)
         {
         }
     }
