@@ -1,5 +1,8 @@
-﻿namespace Moq.AutoMock.Tests
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Moq.AutoMock.Tests
 {
+    [SuppressMessage("Design", "CA1052:Static holder types should be Static or NotInheritable", Justification = "That's kind of the point here")]
     public class WithStatic
     {
         public static string Get()
