@@ -1,5 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
@@ -9,7 +9,7 @@ namespace Moq.AutoMock.Extensions
 {
     internal static class TypeExtensions
     {
-        internal static bool TryCompileGetter(this AutoMocker autoMocker, Type funcType, out Delegate @delegate)
+        internal static bool TryCompileGetter(this AutoMocker autoMocker, Type funcType, [NotNullWhen(true)] out Delegate? @delegate)
         {
             @delegate = null;
             var stInfo = funcType.GetTypeInfo();
