@@ -5,9 +5,9 @@ namespace Moq.AutoMock
 
     internal sealed class RealInstance : IInstance
     {
-        public RealInstance(object value) => Value = value;
+        public RealInstance(object? value) => Value = value;
 
-        public object Value { get; }
+        public object? Value { get; }
         public bool IsMock => false;
     }
 }
