@@ -6,7 +6,7 @@ namespace Moq.AutoMock.Resolvers
     {
         public MockResolutionContext(AutoMocker autoMocker, Type requestType)
             : this(autoMocker, requestType, null) { }
-        public MockResolutionContext(AutoMocker autoMocker, Type requestType, object initialValue)
+        public MockResolutionContext(AutoMocker autoMocker, Type requestType, object? initialValue)
         {
             AutoMocker = autoMocker ?? throw new ArgumentNullException(nameof(autoMocker));
             RequestType = requestType ?? throw new ArgumentNullException(nameof(requestType));
@@ -15,9 +15,9 @@ namespace Moq.AutoMock.Resolvers
 
         public AutoMocker AutoMocker { get; }
         public Type RequestType { get; }
-        public object Value { get; set; }
+        public object? Value { get; set; }
 
-        public void Deconstruct(out AutoMocker autoMocker, out Type type, out object value)
+        public void Deconstruct(out AutoMocker autoMocker, out Type type, out object? value)
         {
             autoMocker = AutoMocker;
             type = RequestType;
